backend/requirements.txt
# Core FastAPI
fastapi
uvicorn[standard]

# Database
sqlalchemy
databases

# Authentication & Security
passlib[bcrypt]
python-jose[cryptography]
python-multipart

# Validation
pydantic[email]
email-validator

# Environment variables
python-dotenv
